+++ gazon-service/src/main/java/com/gitlab/model/Product.java
@@ -5,15 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.NamedAttributeNode;
-import javax.persistence.NamedEntityGraph;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.math.BigDecimal;
 import java.util.Set;
 
@@ -39,9 +31,6 @@ public class Product {
 
     @OneToMany(mappedBy = "someProduct")
     private Set<ProductImage> productImages;
-
-    @OneToMany(mappedBy = "product")
-    private Set<SelectedProduct> selectedProducts;
 
     @OneToMany(mappedBy = "product")
     private Set<Review> review;