+++ gazon-service/src/main/resources/db/changelog/v-1.13/01-add-user-id-column.sql
@@ -1,0 +1,2 @@
+ALTER TABLE selected_product
+    ADD COLUMN user_id BIGINT;