+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - #task_number - short task description
+- #73 - added User field to SelectedProduct
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity